Add initial import feature (Issue 5).
User specifies a ticket to assign to all registration IDs and takes in a CSV file containing the remaining information. User will be shown a list of success/failures after the import is complete.

NOTE: No email is sent yet, but rows that were successfully imported are grouped by email.