Closes issue #17.
The next_id for all passes belonging to the convention that a ticket was
purchased for is now incremented.

So, if two different passes/tickets are available for purchase (say A and
B), purchasing A would increment the next_id value for both A and B.
Purchasing a B would then then be fine.

NOTE: This does not affect manual (admin) registration creations and
registration csv import.

Consideration on improving the flow of ticket purchasing should probably be
considered in the future.